+++ imports/api/works/work.methods.js
@@ -1,5 +1,6 @@
 import { Meteor } from 'meteor/meteor';
-import { getHost } from '../shared';
+import { getHost } from '../@/shared';
+import Works from './work';
 
 Meteor.methods({
   getAllWorks() {
