+++ imports/startup/server/api.js
@@ -16,6 +16,10 @@ import '../../api/pages/server/pages.publications';
 import '../../api/categories/category';
 import '../../api/categories/categories.methods';
 
+import '../../api/works/work';
+import '../../api/works/work.methods';
+import '../../api/works/work.publications';
+
 
 /* old */
 // import '../../api/@/methods/accounts-general';
@@ -26,7 +30,7 @@ import '../../api/@/methods/document-methods';
 import '../../api/@/methods/process-methods';
 // import '../../api/@/methods/profile-methods';
 import '../../api/@/methods/resources-methods';
-import '../../api/@/methods/work-methods';
+// import '../../api/@/methods/work-methods';
 /* old */
