+++ lib/collections.js
@@ -4,7 +4,7 @@ import { Mongo } from 'meteor/mongo';
 
 Activities = new Mongo.Collection('activities');
 Processes = new Mongo.Collection('processes');
-Works = new Mongo.Collection('works');
+// Works = new Mongo.Collection('works');
 // Pages = new Mongo.Collection('pages');
 
 Resources = new Mongo.Collection('resources');
